moving initGameBoard function to helper.js, better ids & classes naming
