+++ index.html
@@ -51,8 +51,8 @@
       <button class="keyboard-btn">Enter</button>
     </div>
   </div>
-  <div class="open-hint-section-btn">GET A HINT</div>
-  <section id="hint-section" class="hidden">
+  <div class="show-hint-container-btn">GET A HINT</div>
+  <section id="hint-container" class="hidden">
     <input type="text" id="hint-input" placeholder="Enter letters, min 1 and max 5"
       title="Enter the letters that you already know are present in the word" />
     <h3 class="hint-title">Results:</h3>
